CUDA 3.2 adaptation and new benchmark tests

+++ src/cux/cux.cpp
@@ -1,4 +1,4 @@
-/***************************************************************************
+/*
  *   Copyright (C) 2004 by Mario Juric                                     *
  *   mjuric@astro.Princeton.EDU                                            *
  *                                                                         *
@@ -438,7 +438,7 @@ bool cux_init()
 
 #if !__DEVICE_EMULATION__
 	// Memory info
-	unsigned free = 0, total = 0;
+	size_t free = 0, total = 0;
 	cuxErrCheck( (cudaError)cuMemGetInfo(&free, &total) );
 //	MLOG(verb2) << "Device memory (free, total): " << free / (1<<20) << "M, " << total / (1<<20) << "M" << "\n";
 #endif
